+++ src/common/types/character-properties.ts
@@ -59,7 +59,15 @@ export type PassiveTalent =
   | `${'25'}% ${'time consumption reduction' | 'more rewards'} when on expeditions in ${TeyvatRegion}.`
   | 'Party members will not startle Crystalflies and certain other animals.'
   | `Party members will not startle animals who produce: Fowl, Raw or Chilled meat.`
-  
+
+  // Traveler
+  | `CRIT Rate +10%. When Pyro/Hydro/CryoElectro attacks from nearby party member hit opponents, gain 1 stack of Blade of the Dawn BReeze. Each Elemental type grants 1 stack. AFter obtaining 2+, performing CA consumes all stacks to transform into CA: Whirlwind: Each CA deals Anemo DMG. +60% of Traveler's ATK. Every consumption, Blade Wind summoned dealing 50% of Traveler's ATK - considered CA damage. 15s CD.`
+  | `DEF +20%. When members protected by shield and shield is replaced or Geo Constructs are created, Traveler receives 1 stack of Blade of Archaic Petra of 3 stacks. 1 stack/s. After 3, performing CA consumes all stacks to transform into CA: Rockfell. Each CA deals Geo DMG +120% of Traveler's ATK. After hit of opponent, +20% shield strength for party members for 15s. 15s CD,`
+  | `Energy Recharge +20%. When members burst/ult, Traveler receives 1 stack of Blade of Resounding Thunder of 3 stacks. After 3, performing CA consumes all stacks to transform into CA: Detonate. Each CA deals Electro DMG +100% of Traveler's ATK. 3s after hitting opponent, 1 additional lightning strike triggered, dealing Electro DMG of 200% Traveler's ATK. 1 Abundance Amulet created. DMG is considered CA DMG. 15s CD.`
+  | `Elemental Master +60. When Dendro/Bountiful Cores or Seeds of Deceit appears on field or party obtains Verdant Dew, Traveler gains 1 stack of Blade of Verdant Virids. These 2 methods each grant +1 stack/4s of 3. After 3, performing CA consumes all stacks to transform into CA: Verdessence. Each CA deals Dendro DMG +80% of Traveler's ATK. Two Vinecores created, which explode after 4/8s respectively, dealing Dendro DMG of 120% of Traveler's ATk - considered CA DMG. 15s CD.`
+  | `HP +20%. When party members' HP change by 5%, Traveler receives 1 stack of Blade of Many Waters every 4s of 3 stacks. After 3, performing CA consumes all stacks to transform into CA: Tidebound. Each CA deals Hydro DMG +150% of Traveler's ATK. When Traveler's HP > 50%, 10% of max HP consumed to crease DMG inflicted by CA by 100% of Traveler's ATK. When HP < 50%, one instance of healing = 25% of max HP restored upon hitting an enemy. 15s CD.`
+  | `ATK +20%. When party members trigger Nightsoul burst, Traveler receives 1 stack of Blade of Sacred Flame of 2 stacks. After 3, performing CA consumes all stacks to transform into CA: Inferno. Each CA deals Nightsoul-aligned Pyro DMG +200% of Traveler's ATK. 15s CD.`
+
   // Special
   | `Increase your own party members' Normal Attack level by 1.`
   | `${'20'}% of double catch when fishing in ${TeyvatRegion}.`